Actualizo Auto.java

+++ Clases/src/com/krakedev/Auto.java
@@ -3,5 +3,5 @@ package com.krakedev;
 public class Auto {
 	public String marca;
 	public int anio;
-	public float precio; 
+	public double precio; 
 }
